+++ CHANGES.rst
@@ -9,6 +9,7 @@
 - Fixed bucket S3 path for Datomic db backup.
 - datomic-free does not support direct `s3` upload -
   work around that with local back and upload via AWS APIs.
+- Updates to reflect switch to non-ephemeral instance.
 
 
 0.17 (2016-06-27)